replace sprintf with com_sprintf for QW.

+++ qw/common.h
@@ -289,8 +289,8 @@ extern	qboolean		standard_quake, rogue, hipnotic;
 char *Info_ValueForKey (char *s, char *key);
 void Info_RemoveKey (char *s, char *key);
 void Info_RemovePrefixedKeys (char *start, char prefix);
-void Info_SetValueForKey (char *s, char *key, char *value, int maxsize);
-void Info_SetValueForStarKey (char *s, char *key, const char *value, int maxsize);
+void Info_SetValueForKey (char *s, char *key, char *value, size_t maxsize);
+void Info_SetValueForStarKey (char *s, char *key, const char *value, size_t maxsize);
 void Info_Print (char *s);
 
 unsigned Com_BlockChecksum (void *buffer, int length);
